+++ src/config/constants/args.rs
@@ -1,11 +1,11 @@
 pub struct Args {
-    pub width: u32,
-    pub height: u32,
-    pub mines: u32,
+    pub width: usize,
+    pub height: usize,
+    pub mines: usize,
 }
 
 impl Args {
-    pub fn new(width: u32, height: u32, mines: u32) -> Result<Args, String> {
+    pub fn new(width: usize, height: usize, mines: usize) -> Result<Args, String> {
         if width < 1 {
             return Err("Width must be greater than 0".to_string());
         }
